+++ app/src/main/java/org/sopt/dosopttemplate/AuthService.kt
@@ -1,0 +1,58 @@
+package org.sopt.dosopttemplate
+
+import android.util.Log
+import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
+import kotlinx.serialization.json.Json
+import okhttp3.Interceptor
+import okhttp3.MediaType.Companion.toMediaType
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Call
+import retrofit2.Retrofit
+import retrofit2.http.Body
+import retrofit2.http.GET
+import retrofit2.http.POST
+import retrofit2.http.Path
+
+interface AuthService {
+    @POST("api/v1/members/sign-in")
+    fun signIn(
+        @Body request: RequestSignInDto,
+    ): Call<ResponseSignInDto>
+
+    @POST("api/v1/members")
+    fun signUp(
+        @Body request: RequestSignUpDto,
+    ): Call<Unit>
+
+    @GET("api/v1/members/{memberId}")
+    fun getUserInfo(
+        @Path("memberId") id: Int,
+    ): Call<ResponseGetUserInfoDto>
+}
+
+object AuthApiFactory {
+    private const val AUTH_BASE_URL = BuildConfig.AUTH_BASE_URL
+
+    private fun getLogOkHttpClient(): Interceptor {
+        val loggingInterceptor = HttpLoggingInterceptor { message ->
+            Log.d("Retrofit2", "CONNECTION INFO -> $message")
+        }
+        loggingInterceptor.level = HttpLoggingInterceptor.Level.BODY
+        return loggingInterceptor
+    }
+
+    private val okHttpClient = OkHttpClient.Builder()
+        .addInterceptor(getLogOkHttpClient())
+        .build()
+
+    val authRetrofit: Retrofit by lazy {
+        Retrofit.Builder()
+            .baseUrl(AUTH_BASE_URL)
+            .client(okHttpClient)
+            .addConverterFactory(Json.asConverterFactory("application/json".toMediaType()))
+            .build()
+    }
+
+    inline fun <reified T> create(): T = authRetrofit.create<T>(T::class.java)
+}
