+++ app/src/main/java/org/sopt/dosopttemplate/FriendType.kt
@@ -1,18 +1,16 @@
 package org.sopt.dosopttemplate
 
-import java.time.LocalDate
-
 
 sealed class Person {
     abstract val profileImage: String
     abstract val userId: String
     abstract val name: String
-    abstract val self_description: String
+    abstract val description: String
 }
 
 data class Friend(
     override val profileImage: String,
     override val userId: String,
     override val name: String,
-    override val self_description: String,
+    override val description: String,
 ) : Person()
