+++ app/src/main/java/org/sopt/dosopttemplate/ServicePool.kt
@@ -1,0 +1,6 @@
+package org.sopt.dosopttemplate
+
+object ServicePool {
+    val authService = AuthApiFactory.create<AuthService>()
+    val friendService = FriendApiFactory.create<FriendService>()
+}
