+++ app/src/main/java/org/sopt/dosopttemplate/SignUpActivity.kt
@@ -6,21 +6,11 @@ import androidx.activity.OnBackPressedCallback
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.snackbar.Snackbar
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
 import org.sopt.dosopttemplate.ServicePool.authService
 import org.sopt.dosopttemplate.databinding.ActivitySignupBinding
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-
-@Serializable
-data class UserInfo(
-    @SerialName("profileImage") val profileImage: String,
-    @SerialName("mbti") val mbti: String,
-    @SerialName("birthday") val birthday: String,
-    @SerialName("self_description") val self_description: String
-)
 
 
 enum class MBTIType {
@@ -29,9 +19,7 @@ enum class MBTIType {
 
 class SignupActivity : AppCompatActivity() {
     private lateinit var binding: ActivitySignupBinding
-    private val sharedPreferences by lazy {
-        getSharedPreferences(SharedPreferencesKeys.USER_INFO, MODE_PRIVATE)
-    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivitySignupBinding.inflate(layoutInflater)
@@ -56,8 +44,13 @@ class SignupActivity : AppCompatActivity() {
                 !isUserNameValid(userName) -> getString(R.string.user_id_error)
                 !isPasswordValid(password) -> getString(R.string.password_error)
                 !isNickNameValid(nickName) -> getString(R.string.nickname_error)
-                !isMBTIValid(binding.etSignUpInputMBTI.text.toString().uppercase()) -> getString(R.string.mbti_error)
-                binding.etSignUpInputMBTI.text.toString().uppercase() == "SEXY" -> getString(R.string.sexy_error)
+                !isMBTIValid(
+                    binding.etSignUpInputMBTI.text.toString().uppercase()
+                ) -> getString(R.string.mbti_error)
+
+                binding.etSignUpInputMBTI.text.toString()
+                    .uppercase() == "SEXY" -> getString(R.string.sexy_error)
+
                 else -> null
             }
 
@@ -74,22 +67,30 @@ class SignupActivity : AppCompatActivity() {
                                 201 -> {
                                     val userInfo = UserInfo(
                                         profileImage = "https://avatars.githubusercontent.com/u/127238018?v=4", // 임시 데이터
-                                        mbti = binding.etSignUpInputMBTI.text.toString().uppercase(),
+                                        mbti = binding.etSignUpInputMBTI.text.toString()
+                                            .uppercase(),
                                         birthday = defaultUserInfo.birthday, // 임시 데이터
                                         self_description = binding.root.context.getString(R.string.test_text) // 임시 데이터
                                     )
 
                                     val userInfoJson = userInfo.toJson()
-                                    userInfoJson.saveAsJsonFile("user_info.json", this@SignupActivity)
+                                    userInfoJson.saveAsJsonFile(
+                                        "user_info.json",
+                                        this@SignupActivity
+                                    )
 
                                     binding.root.showSnackbar(getString(R.string.signup_success_message))
-                                    val intent = Intent(this@SignupActivity, SigninActivity::class.java)
+                                    val intent =
+                                        Intent(this@SignupActivity, SigninActivity::class.java)
                                     startActivity(intent)
                                     finish()
                                 }
+
                                 400 -> {
                                     val errorResponse = response.errorBody()?.string()
-                                    binding.root.showSnackbar(errorResponse ?: getString(R.string.signup_failed))
+                                    binding.root.showSnackbar(
+                                        errorResponse ?: getString(R.string.signup_failed)
+                                    )
                                 }
                             }
                         }
@@ -97,7 +98,7 @@ class SignupActivity : AppCompatActivity() {
                         override fun onFailure(call: Call<Unit>, t: Throwable) {
                             binding.root.showSnackbar(getString(R.string.server_error))
                         }
-                })
+                    })
             }
         }
     }
